profile-app/requirements.txt
Flask==2.3.4
mysql-connector-python==8.1.0
Werkzeug==2.3.7


